Completed crud operations item and customer classes

+++ Thogakade/src/main/java/org/example/dao/custom/impl/CustomerDAOImpl.java
@@ -4,76 +4,53 @@ package org.example.dao.custom.impl;
 import com.sun.xml.bind.v2.model.core.ID;
 import org.example.dao.custom.CustomerDAO;
 import org.example.entity.Customer;
-import org.example.entity.Order;
 import org.example.util.SessionFactoryConfig;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
+import org.hibernate.query.Query;
 
-import java.util.ArrayList;
 import java.util.List;
+
 
 public class CustomerDAOImpl implements CustomerDAO {
 
     @Override
-    public boolean save(Customer customer) {
-        try (Session session = SessionFactoryConfig.getInstance().getSession()) {
-            Transaction transaction = session.beginTransaction();
-            try {
-                session.save(customer);
-                transaction.commit();
-                return true;
-            } catch (Exception e) {
-                if (transaction != null) transaction.rollback();
-                return false;
-            }
-        }
-    }
-
-    @Override
-    public Customer findById(ID id) {
-        try (Session session = SessionFactoryConfig.getInstance().getSession()) {
-            return session.get(Customer.class, id);
-        }
-    }
-
-    @Override
-    public boolean update(Customer customer) {
-        try(Session session=SessionFactoryConfig.getInstance().getSession()){
-            Transaction transaction=session.beginTransaction();
-
-            try{
-                session.save(customer);
-                transaction.commit();
-                return true;
-            }catch (Exception e){
-                if(transaction!=null) transaction.rollback();
-                return false;
-            }
-        }
+    public boolean save(Customer customer, Session session) throws RuntimeException {
+        session.save(customer);
+        return true;
     }
 
 
-
     @Override
-    public boolean delete(String id) {
-        try (Session session = SessionFactoryConfig.getInstance().getSession()) {
-            Transaction transaction = session.beginTransaction();
-            try {
-                Customer customer = session.load(Customer.class, id);
-                session.delete(customer);
-                transaction.commit();
-                return true;
-            } catch (Exception e) {
-                if (transaction != null) transaction.rollback();
-                return false;
-            }
-        }
+    public boolean update(Customer customer, Session session) throws RuntimeException {
+        session.update(customer);
+        return true;
     }
 
     @Override
-    public List<Customer> getAll() {
-        try (Session session = SessionFactoryConfig.getInstance().getSession()) {
-            return session.createQuery("FROM Customer", Customer.class).list();
+    public Boolean delete(String id, Session session) throws RuntimeException {
+        Customer customer = session.get(Customer.class, id);
+        if (customer == null) {
+            throw new RuntimeException("Customer with ID " + id + " not found, cannot delete.");
         }
+        session.delete(customer);
+        System.out.println("Customer successfully deleted in DAO.");
+        return true;
+    }
+
+    @Override
+    public List<Customer> getAll(Session session) {
+        String sql = "FROM Customer ";
+        Query query = session.createQuery(sql);
+        List<Customer> list = query.list();
+        return list;
+    }
+
+
+    @Override
+    public List<Customer> searchCustomerByText(String text, Session session) {
+        Query query = session.createQuery("FROM Customer  WHERE name LIKE '%" + text + "%'");
+        List<Customer> list = query.list();
+        return list;
     }
 }
